+++ test/exactly_lib_test/test_case_utils/line_matcher/test_resources/integration_check.py
@@ -28,6 +28,12 @@ CHECKER = integration_check.IntegrationChecker(
     False,
 )
 
+CHECKER__PARSE_SIMPLE = integration_check.IntegrationChecker(
+    parse_line_matcher.parsers(True).simple,
+    MatcherPropertiesConfiguration(),
+    False,
+)
+
 
 def check(put: unittest.TestCase,
           source: ParseSource,